Data Pipeline V2: Move Face Detection to the new DataModule (#1041)
